added

+++ node-server/deletedata/userdata.js
@@ -6,9 +6,12 @@ const router = express.Router();
 
     router.delete('/:id',async function(req,res){
         console.log("deleting users data..");
-       await user.db1.deleteOne({userid:req.params.id},function(err){ 
+       await user.db1.deleteOne({userid:req.params.id} || {userid:null},function(err){ 
             console.log(`data having id ${req.params.id} is now delete`)
          
         })        
     })
+
+
+
 module.exports = router;
